requirements.txt
asgiref==3.3.4
cached-property==1.5.2
chardet==4.0.0
csv-reader==1.2.0
Django==3.2.3
django-filter==2.4.0
django-url-filter==0.3.15
djangorestframework==3.12.4
enum-compat==0.0.3
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
